Some bugfixing done, a bunch of new features added. The item page activity is now entirely card-based, and shows the description (not the entire page) in a WebView contained in a CardView, as opposed to trying to parse the HTML myself.
An image gallery for the item activity is also well under way, also contained within a CardView and currently showing some repeated placeholder images. I still need to make it scrape the real images from the page, and show an enlarged version somehow on request (i.e. a tap on the image).